# parser_syslog: Support configured time_format for rfc3164 string parser